feat(Arkanun 2e)
Ajustadas regras conforme referências do Arkanun 2e.

+++ System.Daemon/Interface/IRolagem.cs
@@ -3,6 +3,6 @@ namespace System.Daemon.Interface
     public interface IRolagem
     {
         float RolarD100();
-        int RolarDado(byte quantidade, byte faces, int modificador);
+        byte RolarDado(byte quantidade, byte faces, int modificador);
     }
 }
